code-review/code_review_env/server/requirements.txt
openenv-core[core]>=0.2.3
fastapi>=0.115.0
uvicorn>=0.30.0
